Fixed simple api tests, refactored into classes
